feat: moved DataGroups to its own dir
