+++ src/dataDisplay/dataGroups/dataGroup.jsx
@@ -2,10 +2,10 @@ import _ from 'lodash';
 import ClassNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
-import ActivityIndicator from '../atoms/activityIndicator';
-import Header from '../atoms/header';
-import Icon from './icon';
-import List from './list';
+import ActivityIndicator from '../../atoms/activityIndicator';
+import Header from '../../atoms/header';
+import Icon from '../icon';
+import List from '../list';
 import DataGroupExpandedRow from './dataGroupExpandedRow';
 import DataGroupRow from './dataGroupRow';
 import { groupPropTypes } from './dataGroupsPropTypes';
