+++ src/dataDisplay/dataGroups/dataGroupExpandedRow.jsx
@@ -2,9 +2,9 @@ import _ from 'lodash';
 import ClassNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
-import Button from '../inputs/button';
-import Header from '../atoms/header';
-import Icon from './icon';
+import Button from '../../inputs/button';
+import Header from '../../atoms/header';
+import Icon from '../icon';
 import { rowPropTypes } from './dataGroupsPropTypes';
 
 const propTypes = {
