+++ src/dataDisplay/dataGroups/dataGroupRow.jsx
@@ -2,8 +2,8 @@ import _ from 'lodash';
 import ClassNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
-import Header from '../atoms/header';
-import Icon from './icon';
+import Header from '../../atoms/header';
+import Icon from '../icon';
 import { rowPropTypes } from './dataGroupsPropTypes';
 
 const propTypes = {
