+++ src/style.scss
@@ -10,7 +10,7 @@
 
 // NEW: Data Display
 @import './dataDisplay/banner';
-@import './dataDisplay/dataGroups';
+@import './dataDisplay/dataGroups/dataGroups';
 @import './dataDisplay/divider';
 @import './dataDisplay/icon/icon';
 @import './dataDisplay/image';
